chore: check basic permissions

+++ src/main/java/com/grash/controller/AssetController.java
@@ -5,6 +5,7 @@ import com.grash.dto.SuccessResponse;
 import com.grash.exception.CustomException;
 import com.grash.model.Asset;
 import com.grash.model.User;
+import com.grash.model.enums.BasicPermission;
 import com.grash.model.enums.RoleType;
 import com.grash.service.AssetService;
 import com.grash.service.UserService;
@@ -105,7 +106,8 @@ public class AssetController {
         Optional<Asset> optionalAsset = assetService.findById(id);
         if (optionalAsset.isPresent()) {
             Asset savedAsset = optionalAsset.get();
-            if (assetService.hasAccess(user, savedAsset)) {
+            if (assetService.hasAccess(user, savedAsset)
+                    && user.getRole().getPermissions().contains(BasicPermission.DELETE_ASSETS)) {
                 assetService.delete(id);
                 return new ResponseEntity(new SuccessResponse(true, "Deleted successfully"),
                         HttpStatus.OK);
