+++ src/main/java/com/grash/controller/MultiPartsController.java
@@ -5,6 +5,7 @@ import com.grash.dto.SuccessResponse;
 import com.grash.exception.CustomException;
 import com.grash.model.MultiParts;
 import com.grash.model.User;
+import com.grash.model.enums.BasicPermission;
 import com.grash.model.enums.RoleType;
 import com.grash.service.MultiPartsService;
 import com.grash.service.UserService;
@@ -105,7 +106,8 @@ public class MultiPartsController {
         Optional<MultiParts> optionalMultiParts = multiPartsService.findById(id);
         if (optionalMultiParts.isPresent()) {
             MultiParts savedMultiParts = optionalMultiParts.get();
-            if (multiPartsService.hasAccess(user, savedMultiParts)) {
+            if (multiPartsService.hasAccess(user, savedMultiParts)
+                    && user.getRole().getPermissions().contains(BasicPermission.DELETE_PARTS_AND_MULTI_PARTS)) {
                 multiPartsService.delete(id);
                 return new ResponseEntity(new SuccessResponse(true, "Deleted successfully"),
                         HttpStatus.OK);
